CIV-13701 Move handler to claimant package
